+++ weblate/checks/models.py
@@ -74,6 +74,7 @@ class WeblateChecksConf(AppConf):
         "weblate.checks.format.JavaFormatCheck",
         "weblate.checks.format.JavaMessageFormatCheck",
         "weblate.checks.format.PercentPlaceholdersCheck",
+        "weblate.checks.format.VueFormattingCheck",
         "weblate.checks.format.I18NextInterpolationCheck",
         "weblate.checks.format.ESTemplateLiteralsCheck",
         "weblate.checks.angularjs.AngularJSInterpolationCheck",